рефакторинг страницы новостей
