+++ lib/feature/presentation/ui/news_page/articles/single_article/single_news_article_mobile.dart
@@ -1,0 +1,121 @@
+import 'package:cportal_flutter/app_config.dart';
+import 'package:cportal_flutter/common/custom_theme.dart';
+import 'package:cportal_flutter/common/util/padding.dart';
+import 'package:cportal_flutter/feature/domain/entities/article_entity.dart';
+import 'package:cportal_flutter/feature/presentation/ui/news_page/widgets/news_template.dart';
+import 'package:cportal_flutter/feature/presentation/ui/widgets/desktop_menu.dart';
+import 'package:cportal_flutter/feature/presentation/ui/widgets/news_main_mobile.dart';
+import 'package:extended_image/extended_image.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:intl/intl.dart';
+import 'package:responsive_framework/responsive_wrapper.dart';
+
+class SingleNewsArticleMobile extends StatelessWidget {
+  final ArticleEntity article;
+  final List<ArticleEntity> articles;
+
+  const SingleNewsArticleMobile({
+    Key? key,
+    required this.article,
+    required this.articles,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
+    final DateFormat outputFormat = DateFormat('d MMMM y, H:m', 'ru');
+
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        if (ResponsiveWrapper.of(context).isLargerThan(MOBILE))
+          DesktopMenu(
+            currentIndex: 1,
+            onChange: (index) {
+              changePage(context, index);
+            },
+          ),
+        Expanded(
+          child: SafeArea(
+            child: CustomScrollView(
+              physics: const BouncingScrollPhysics(),
+              slivers: [
+                SliverAppBar(
+                  backgroundColor: theme.background,
+                  expandedHeight: 176,
+                  automaticallyImplyLeading: false,
+                  leading: IconButton(
+                    highlightColor: Colors.transparent,
+                    splashColor: Colors.transparent,
+                    enableFeedback: false,
+                    icon: const Icon(Icons.arrow_back),
+                    iconSize: 24,
+                    onPressed: () => GoRouter.of(context).pop(),
+                  ),
+                  flexibleSpace: FlexibleSpaceBar(
+                    background: ExtendedImage.network(
+                      '${AppConfig.apiUri}/images/${article.image}',
+                      fit: BoxFit.cover,
+                      cache: true,
+                    ),
+                  ),
+                ),
+
+                /// Колонка с контентом статьи.
+                SliverToBoxAdapter(
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(vertical: 16),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Padding(
+                          padding: getHorizontalPadding(context),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                article.header,
+                                style: theme.textTheme.px22,
+                              ),
+                              const SizedBox(height: 4),
+                              Text(
+                                outputFormat.format(
+                                  article.date,
+                                ),
+                                style: theme.textTheme.px12,
+                              ),
+                              const SizedBox(height: 20),
+                              Column(
+                                children: [
+                                  ...List.generate(
+                                    article.content.length,
+                                    (index) {
+                                      return NewsTemplate.factory(
+                                        context,
+                                        article.content[index],
+                                      );
+                                    },
+                                  ),
+                                ],
+                              ),
+                              const SizedBox(height: 20),
+                            ],
+                          ),
+                        ),
+                        NewsMainMobile(
+                          articles: articles,
+                          currentArticle: article,
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
